# Generate point clouds on device