--- a/benchmarks/utils/ArborXBenchmark_PointClouds.hpp
+++ b/benchmarks/utils/ArborXBenchmark_PointClouds.hpp
@@ -17,6 +17,7 @@
 #include <ArborX_GeometryTraits.hpp>
 
 #include <Kokkos_Core.hpp>
+#include <Kokkos_Random.hpp>
 
 #include <fstream>
 #include <random>
@@ -49,160 +50,217 @@ inline PointCloudType to_point_cloud_enum(std::string const &str)
 namespace Details
 {
 
-template <class Point, typename... ViewProperties>
-void filledBoxCloud(double const half_edge,
-                    Kokkos::View<Point *, ViewProperties...> random_points)
+template <typename ExecutionSpace, typename Points>
+void filledBoxCloud(ExecutionSpace const &exec, double const half_edge,
+                    Points &random_points)
 {
-  namespace KokkosExt = ArborX::Details::KokkosExt;
-  static_assert(
-      KokkosExt::is_accessible_from_host<decltype(random_points)>::value,
-      "The View should be accessible on the Host");
-  std::uniform_real_distribution<double> distribution(-half_edge, half_edge);
-  std::default_random_engine generator;
-  auto random = [&distribution, &generator]() {
-    return distribution(generator);
-  };
+  using Point = typename Points::value_type;
+  constexpr auto DIM = ArborX::GeometryTraits::dimension_v<Point>;
+  using Coordinate = ArborX::GeometryTraits::coordinate_type_t<Point>;
+
+  using GeneratorPool = Kokkos::Random_XorShift1024_Pool<ExecutionSpace>;
+  using GeneratorType = typename GeneratorPool::generator_type;
+  constexpr unsigned int batch_size = 8;
+
+  GeneratorPool random_pool(0);
   unsigned int const n = random_points.extent(0);
-  constexpr auto DIM = ArborX::GeometryTraits::dimension<Point>::value;
-  for (unsigned int i = 0; i < n; ++i)
-    for (int d = 0; d < DIM; ++d)
-      random_points(i)[d] = random();
+  Kokkos::parallel_for(
+      "ArborXBenchmark::filledBoxCloud::generate",
+      Kokkos::RangePolicy<ExecutionSpace>(exec, 0, n / batch_size),
+      KOKKOS_LAMBDA(int i) {
+        auto generator = random_pool.get_state();
+        auto random = [&generator, half_edge]() {
+          return Kokkos::rand<GeneratorType, Coordinate>::draw(
+              generator, -half_edge, half_edge);
+        };
+
+        auto begin = i * batch_size;
+        auto end = Kokkos::min((i + 1) * batch_size, n);
+        for (unsigned int k = begin; k < end; ++k)
+          for (int d = 0; d < DIM; ++d)
+            random_points(k)[d] = random();
+
+        random_pool.free_state(generator);
+      });
 }
 
-template <class Point, typename... ViewProperties>
-void hollowBoxCloud(double const half_edge,
-                    Kokkos::View<Point *, ViewProperties...> random_points)
+template <typename ExecutionSpace, typename Points>
+void hollowBoxCloud(ExecutionSpace const &exec, double const half_edge,
+                    Points &random_points)
 {
-  namespace KokkosExt = ArborX::Details::KokkosExt;
-  static_assert(
-      KokkosExt::is_accessible_from_host<decltype(random_points)>::value,
-      "The View should be accessible on the Host");
-  std::uniform_real_distribution<double> distribution(-half_edge, half_edge);
-  std::default_random_engine generator;
-  auto random = [&distribution, &generator]() {
-    return distribution(generator);
-  };
+  using Point = typename Points::value_type;
+  constexpr auto DIM = ArborX::GeometryTraits::dimension_v<Point>;
+  using Coordinate = ArborX::GeometryTraits::coordinate_type_t<Point>;
+
+  using GeneratorPool = Kokkos::Random_XorShift1024_Pool<ExecutionSpace>;
+  using GeneratorType = typename GeneratorPool::generator_type;
+  constexpr unsigned int batch_size = 8;
+
+  GeneratorPool random_pool(0);
   unsigned int const n = random_points.extent(0);
-  constexpr auto DIM = ArborX::GeometryTraits::dimension<Point>::value;
   // Points are cyclically placed on the faces of a box
-  constexpr int num_faces = 2 * DIM;
-  for (int axis = 0; axis < DIM; ++axis)
-  {
-    for (unsigned int i = 2 * axis; i < n; i += num_faces)
-      for (int d = 0; d < DIM; ++d)
-        random_points(i)[d] = (d != axis ? random() : -half_edge);
+  Kokkos::parallel_for(
+      "ArborXBenchmark::hollowBoxCloud::generate",
+      Kokkos::RangePolicy<ExecutionSpace>(exec, 0, n / batch_size),
+      KOKKOS_LAMBDA(int i) {
+        auto generator = random_pool.get_state();
+        auto random = [&generator, half_edge]() {
+          return Kokkos::rand<GeneratorType, Coordinate>::draw(
+              generator, -half_edge, half_edge);
+        };
 
-    for (unsigned int i = 2 * axis + 1; i < n; i += num_faces)
-      for (int d = 0; d < DIM; ++d)
-        random_points(i)[d] = (d != axis ? random() : +half_edge);
-  }
+        auto begin = i * batch_size;
+        auto end = Kokkos::min((i + 1) * batch_size, n);
+        for (unsigned int k = begin; k < end; ++k)
+        {
+          // For 3D, the order is
+          // Indices: 0  1  2  3  4  5  6  7  8  9
+          // Axes   : 0- 0+ 1- 1+ 2- 2+ 0- 0+ 1- 1+
+          int axis = (k / 2) % DIM;
+          for (int d = 0; d < DIM; ++d)
+          {
+            if (d != axis)
+              random_points(k)[d] = random();
+            else if (k % 2 == 0)
+              random_points(k)[d] = -half_edge;
+            else
+              random_points(k)[d] = half_edge;
+          }
+        }
+
+        random_pool.free_state(generator);
+      });
 }
 
-template <class Point, typename... ViewProperties>
-void filledSphereCloud(double const radius,
-                       Kokkos::View<Point *, ViewProperties...> random_points)
+template <typename ExecutionSpace, typename Points>
+void filledSphereCloud(ExecutionSpace const &exec, double const radius,
+                       Points &random_points)
 {
-  namespace KokkosExt = ArborX::Details::KokkosExt;
-  static_assert(
-      KokkosExt::is_accessible_from_host<decltype(random_points)>::value,
-      "The View should be accessible on the Host");
-  std::default_random_engine generator;
+  using Point = typename Points::value_type;
+  constexpr auto DIM = ArborX::GeometryTraits::dimension_v<Point>;
+  using Coordinate = ArborX::GeometryTraits::coordinate_type_t<Point>;
 
-  std::uniform_real_distribution<double> distribution(-radius, radius);
-  auto random = [&distribution, &generator]() {
-    return distribution(generator);
-  };
+  using GeneratorPool = Kokkos::Random_XorShift1024_Pool<ExecutionSpace>;
+  using GeneratorType = typename GeneratorPool::generator_type;
+  constexpr unsigned int batch_size = 8;
 
+  GeneratorPool random_pool(0);
   unsigned int const n = random_points.extent(0);
-  constexpr auto DIM = ArborX::GeometryTraits::dimension<Point>::value;
-  for (unsigned int i = 0; i < n; ++i)
-  {
-    bool point_accepted = false;
-    while (!point_accepted)
-    {
-      Point p;
-      double norm = 0.f;
-      for (int d = 0; d < DIM; ++d)
-      {
-        double v = random();
-        p[d] = v;
-        norm += v * v;
-      }
-      norm = std::sqrt(norm);
+  Kokkos::parallel_for(
+      "ArborXBenchmark::filledSphereCloud::generate",
+      Kokkos::RangePolicy<ExecutionSpace>(exec, 0, n / batch_size),
+      KOKKOS_LAMBDA(int i) {
+        auto generator = random_pool.get_state();
+        auto random = [&generator, radius]() {
+          return Kokkos::rand<GeneratorType, Coordinate>::draw(generator,
+                                                               -radius, radius);
+        };
 
-      // Only accept points that are in the sphere
-      if (norm <= radius)
-      {
-        random_points(i) = p;
-        point_accepted = true;
-      }
-    }
-  }
+        auto begin = i * batch_size;
+        auto end = Kokkos::min((i + 1) * batch_size, n);
+        for (unsigned int k = begin; k < end; ++k)
+        {
+          do
+          {
+            Point p;
+            Coordinate norm = 0;
+            for (int d = 0; d < DIM; ++d)
+            {
+              p[d] = random();
+              norm += p[d] * p[d];
+            }
+            norm = Kokkos::sqrt(norm);
+
+            // Only accept points that are in the sphere
+            if (norm <= radius)
+            {
+              random_points(k) = p;
+              break;
+            }
+          } while (true);
+        }
+
+        random_pool.free_state(generator);
+      });
 }
 
-template <class Point, typename... ViewProperties>
-void hollowSphereCloud(double const radius,
-                       Kokkos::View<Point *, ViewProperties...> random_points)
+template <typename ExecutionSpace, typename Points>
+void hollowSphereCloud(ExecutionSpace const &exec, double const radius,
+                       Points &random_points)
 {
-  namespace KokkosExt = ArborX::Details::KokkosExt;
-  static_assert(
-      KokkosExt::is_accessible_from_host<decltype(random_points)>::value,
-      "The View should be accessible on the Host");
-  std::default_random_engine generator;
+  using Point = typename Points::value_type;
+  constexpr auto DIM = ArborX::GeometryTraits::dimension_v<Point>;
+  using Coordinate = ArborX::GeometryTraits::coordinate_type_t<Point>;
 
-  std::normal_distribution<float> distribution(0.f, 1.f);
-  auto random = [&distribution, &generator]() {
-    return distribution(generator);
-  };
+  using GeneratorPool = Kokkos::Random_XorShift1024_Pool<ExecutionSpace>;
+  using GeneratorType = typename GeneratorPool::generator_type;
+  constexpr unsigned int batch_size = 8;
 
+  GeneratorPool random_pool(0);
   unsigned int const n = random_points.extent(0);
-  constexpr auto DIM = ArborX::GeometryTraits::dimension<Point>::value;
-  for (unsigned int i = 0; i < n; ++i)
-  {
-    double v[DIM];
-    double norm = 0.;
-    for (int d = 0; d < DIM; ++d)
-    {
-      v[d] = random();
-      norm += v[d] * v[d];
-    }
-    norm = std::sqrt(norm);
+  Kokkos::parallel_for(
+      "ArborXBenchmark::hollowSphereCloud::generate",
+      Kokkos::RangePolicy<ExecutionSpace>(exec, 0, n / batch_size),
+      KOKKOS_LAMBDA(int i) {
+        auto generator = random_pool.get_state();
+        auto random = [&generator, radius]() {
+          return Kokkos::rand<GeneratorType, Coordinate>::draw(generator,
+                                                               -radius, radius);
+        };
 
-    for (int d = 0; d < DIM; ++d)
-      random_points(i)[d] = radius * v[d] / norm;
-  }
+        auto begin = i * batch_size;
+        auto end = Kokkos::min((i + 1) * batch_size, n);
+        for (unsigned int k = begin; k < end; ++k)
+        {
+          Point p;
+          Coordinate norm = 0;
+          for (int d = 0; d < DIM; ++d)
+          {
+            p[d] = random();
+            norm += p[d] * p[d];
+          }
+          norm = Kokkos::sqrt(norm);
+
+          for (int d = 0; d < DIM; ++d)
+            random_points(k)[d] = radius * p[d] / norm;
+        }
+
+        random_pool.free_state(generator);
+      });
 }
 
 } // namespace Details
 
-template <class Point, typename DeviceType>
-void generatePointCloud(PointCloudType const point_cloud_type,
-                        double const length,
-                        Kokkos::View<Point *, DeviceType> random_points)
+template <typename ExecutionSpace, typename Points>
+void generatePointCloud(ExecutionSpace const &exec,
+                        PointCloudType const point_cloud_type,
+                        double const length, Points &random_points)
 {
+  static_assert(Kokkos::is_view_v<Points>);
+
+  using Point = typename Points::value_type;
+
   using namespace ArborX::GeometryTraits;
   check_valid_geometry_traits(Point{});
   static_assert(is_point_v<Point>, "ArborX: View must contain point values");
 
-  auto random_points_host = Kokkos::create_mirror_view(random_points);
   switch (point_cloud_type)
   {
   case PointCloudType::filled_box:
-    Details::filledBoxCloud(length, random_points_host);
+    Details::filledBoxCloud(exec, length, random_points);
     break;
   case PointCloudType::hollow_box:
-    Details::hollowBoxCloud(length, random_points_host);
+    Details::hollowBoxCloud(exec, length, random_points);
     break;
   case PointCloudType::filled_sphere:
-    Details::filledSphereCloud(length, random_points_host);
+    Details::filledSphereCloud(exec, length, random_points);
     break;
   case PointCloudType::hollow_sphere:
-    Details::hollowSphereCloud(length, random_points_host);
+    Details::hollowSphereCloud(exec, length, random_points);
     break;
   default:
     throw ArborX::SearchException("not implemented");
   }
-  Kokkos::deep_copy(random_points, random_points_host);
 }
 
 } // namespace ArborXBenchmark