Generating large point clouds (like 100M) on host takes about half a minute (on Frontier), which makes it challenging to run multiple benchmarks. This PR does generation on the device. In addition, this PR fixes point generation for sphere and hollow sphere to produce uniform distributions and work better for high dimensions.

The generation is done in batched mode. I compared generating a certain number of random floats on A100 using different batches using the following driver:
[kokkos_vs_std_random.cpp](https://github.com/user-attachments/files/15537921/kokkos_vs_std_random.txt).

There's not much difference once one goes beyond batch 4:
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
BM_std_random/100                     7.85 us         7.85 us        89068
BM_std_random/10000                   56.6 us         56.6 us        12132
BM_std_random/1000000                 4582 us         4582 us          152
BM_std_random/100000000             457032 us       456937 us            2

BM_kokkos_random/100/1                13.4 us         13.4 us        40553
BM_kokkos_random/10000/1              13.3 us         13.3 us        52073
BM_kokkos_random/1000000/1             146 us          146 us         4831
BM_kokkos_random/100000000/1         13616 us        13616 us           52

BM_kokkos_random/100/4                12.1 us         12.1 us        57673
BM_kokkos_random/10000/4              13.7 us         13.7 us        51221
BM_kokkos_random/1000000/4            68.3 us         68.3 us        10250
BM_kokkos_random/100000000/4          4465 us         4462 us          158

BM_kokkos_random/100/16               13.3 us         13.3 us        52455
BM_kokkos_random/10000/16             16.1 us         16.1 us        43482
BM_kokkos_random/1000000/16           40.4 us         40.4 us        17361
BM_kokkos_random/100000000/16         3964 us         3963 us          177

BM_kokkos_random/100/64               18.3 us         18.3 us        38187
BM_kokkos_random/10000/64             26.6 us         26.6 us        26283
BM_kokkos_random/1000000/64           51.8 us         51.8 us        13491
BM_kokkos_random/100000000/64         5250 us         5249 us          133

BM_kokkos_random/100/256              1.52 us         1.52 us       462197
BM_kokkos_random/10000/256            44.1 us         44.1 us        15880
BM_kokkos_random/1000000/256          69.5 us         69.5 us        10059
BM_kokkos_random/100000000/256        6434 us         6432 us          110

BM_kokkos_random/100/1024             1.51 us         1.51 us       463793
BM_kokkos_random/10000/1024            119 us          119 us         5907
BM_kokkos_random/1000000/1024          230 us          230 us         3052
BM_kokkos_random/100000000/1024       5408 us         5408 us          129

BM_kokkos_random/100/4096             1.51 us         1.51 us       463578
BM_kokkos_random/10000/4096            439 us          439 us         1596
BM_kokkos_random/1000000/4096          872 us          872 us          802
BM_kokkos_random/100000000/4096       4397 us         4396 us          157
```
I chose 8, on the lower end, given that we are going to generate multi-dimensional points, which means `n * DIM`.

I validated the produced clouds by dumping them into VTK and visualizing.