@aprokop 2024-06-03
Here are the performance results:

### master (da4fb8bf) 
[master_da4fb8bf.log](https://github.com/user-attachments/files/15537968/master_da4fb8bf.log)

```
--------------------------------------------------------------------
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
BM_filled_box/100              0.015 ms        0.015 ms        45426
BM_filled_box/1000             0.039 ms        0.039 ms        18009
BM_filled_box/10000            0.270 ms        0.270 ms         2588
BM_filled_box/100000            2.56 ms         2.56 ms          274
BM_filled_box/1000000           24.8 ms         24.8 ms           28
BM_filled_box/10000000           303 ms          303 ms            2
BM_hollow_box/100              0.015 ms        0.015 ms        47096
BM_hollow_box/1000             0.032 ms        0.032 ms        21959
BM_hollow_box/10000            0.202 ms        0.202 ms         3431
BM_hollow_box/100000            1.86 ms         1.86 ms          379
BM_hollow_box/1000000           17.7 ms         17.7 ms           40
BM_hollow_box/10000000           222 ms          222 ms            3
BM_filled_sphere/100           0.018 ms        0.018 ms        39861
BM_filled_sphere/1000          0.067 ms        0.067 ms        10422
BM_filled_sphere/10000         0.682 ms        0.682 ms         1027
BM_filled_sphere/100000         6.73 ms         6.73 ms          104
BM_filled_sphere/1000000        66.3 ms         66.3 ms           10
BM_filled_sphere/10000000        732 ms          732 ms            1
BM_hollow_sphere/100           0.017 ms        0.017 ms        41022
BM_hollow_sphere/1000          0.056 ms        0.056 ms        12549
BM_hollow_sphere/10000         0.464 ms        0.464 ms         1513
BM_hollow_sphere/100000         4.49 ms         4.49 ms          156
BM_hollow_sphere/1000000        44.3 ms         44.3 ms           16
BM_hollow_sphere/10000000        485 ms          485 ms            2
```

### This branch (serial) 
[points_device_serial.log](https://github.com/user-attachments/files/15537972/points_device_serial.log)

```
--------------------------------------------------------------------
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
BM_filled_box/100              0.008 ms        0.008 ms        82958
BM_filled_box/1000             0.024 ms        0.024 ms        28931
BM_filled_box/10000            0.244 ms        0.244 ms         2877
BM_filled_box/100000            2.38 ms         2.38 ms          294
BM_filled_box/1000000           23.8 ms         23.8 ms           29
BM_filled_box/10000000           258 ms          258 ms            2
BM_hollow_box/100              0.008 ms        0.008 ms        84657
BM_hollow_box/1000             0.022 ms        0.022 ms        32099
BM_hollow_box/10000            0.203 ms        0.203 ms         3459
BM_hollow_box/100000            1.98 ms         1.98 ms          353
BM_hollow_box/1000000           19.8 ms         19.8 ms           36
BM_hollow_box/10000000           211 ms          211 ms            3
BM_filled_sphere/100           0.009 ms        0.009 ms        74673
BM_filled_sphere/1000          0.056 ms        0.056 ms        12418
BM_filled_sphere/10000         0.568 ms        0.568 ms         1233
BM_filled_sphere/100000         5.54 ms         5.54 ms          126
BM_filled_sphere/1000000        55.6 ms         55.6 ms           13
BM_filled_sphere/10000000        600 ms          600 ms            1
BM_hollow_sphere/100           0.009 ms        0.009 ms        77972
BM_hollow_sphere/1000          0.028 ms        0.028 ms        24659
BM_hollow_sphere/10000         0.268 ms        0.268 ms         2609
BM_hollow_sphere/100000         2.63 ms         2.62 ms          267
BM_hollow_sphere/1000000        26.2 ms         26.2 ms           27
BM_hollow_sphere/10000000        283 ms          283 ms            2
```

### This branch (device, A100) 
[points_device_a100.log](https://github.com/user-attachments/files/15537974/points_device_a100.log)

```
--------------------------------------------------------------------
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
BM_filled_box/100               28.3 ms         28.3 ms           23
BM_filled_box/1000              28.1 ms         28.1 ms           25
BM_filled_box/10000             28.1 ms         28.1 ms           25
BM_filled_box/100000            28.1 ms         28.1 ms           25
BM_filled_box/1000000           29.8 ms         29.8 ms           24
BM_filled_box/10000000          31.0 ms         31.0 ms           23
BM_hollow_box/100               28.1 ms         28.1 ms           25
BM_hollow_box/1000              28.1 ms         28.1 ms           25
BM_hollow_box/10000             28.0 ms         28.0 ms           25
BM_hollow_box/100000            28.1 ms         28.1 ms           25
BM_hollow_box/1000000           29.4 ms         29.4 ms           23
BM_hollow_box/10000000          31.1 ms         31.1 ms           22
BM_filled_sphere/100            28.0 ms         28.0 ms           25
BM_filled_sphere/1000           28.4 ms         28.4 ms           25
BM_filled_sphere/10000          28.1 ms         28.1 ms           25
BM_filled_sphere/100000         28.1 ms         28.1 ms           25
BM_filled_sphere/1000000        50.4 ms         30.3 ms           24
BM_filled_sphere/10000000       34.6 ms         34.6 ms           20
BM_hollow_sphere/100            28.3 ms         28.2 ms           25
BM_hollow_sphere/1000           28.0 ms         28.0 ms           25
BM_hollow_sphere/10000          28.0 ms         28.0 ms           25
BM_hollow_sphere/100000         28.0 ms         28.0 ms           25
BM_hollow_sphere/1000000        29.3 ms         29.3 ms           24
BM_hollow_sphere/10000000       30.9 ms         30.9 ms           23
```