@aprokop 2024-06-03
As one can see, generating on the device does not really depend on the size, as most of the time is taken by grabbing the state from a random pool. I looked into it, and I still don't have any idea why it is so expensive, when it really should take no time at all. Still, 0.03 seconds is good enough for me.