@aprokop 2024-06-04
Following [this](https://blogs.sas.com/content/iml/2016/04/06/generate-points-uniformly-in-ball.html) blog

> If $Y$ is drawn from the uncorrelated multivariate normal distribution, then $S = Y / \|Y\|$ has the uniform distribution on the unit $d$-sphere. Multiplying $S$ by $U^{1/d}$, where $U$ has the uniform distribution on the unit interval $(0,1)$, creates the uniform distribution in the unit $d$-dimensional ball.